рефакторила код
