+++ js/chooseScreen.js
@@ -14,6 +14,7 @@ export default (data) => {
   };
   if (data.attempt > 0 && data.lives > 0 && data.minute !== 2 && data.seconds !== 59) {
     questionArr[getRandom(2, 1)]();
+    window.initializeCountdown();
   } else {
     if (data.attempt <= 0) {
       result(data);
